Struct parry3d::bounding_volume::SimdAabb

pub struct SimdAabb {
    pub mins: Point<SimdReal>,
    pub maxs: Point<SimdReal>,
}

Four Aabb represented as a single SoA Aabb with SIMD components.

Fields
mins: Point<SimdReal>

The min coordinates of the Aabbs.

maxs: Point<SimdReal>

The max coordinates the Aabbs.

Implementations

impl SimdAabb

pub fn new_invalid() -> Self

An invalid Aabb.

pub fn splat(aabb: Aabb) -> Self

Builds an SIMD aabb composed of four identical aabbs.

pub fn center(&self) -> Point<SimdReal>

The center of all the Aabbs represented by `self``.

pub fn half_extents(&self) -> Vector<SimdReal>

The half-extents of all the Aabbs represented by `self``.

pub fn radius(&self) -> SimdReal

The radius of all the Aabbs represented by `self``.

pub fn transform_by(&self, transform: &Isometry<SimdReal>) -> Self

Return the Aabb of the self transformed by the given isometry.

pub fn scaled(self, scale: &Vector<SimdReal>) -> Self

Returns a scaled version of this Aabb.

pub fn loosen(&mut self, margin: SimdReal)

Enlarges this bounding volume by the given margin.

pub fn dilate_by_factor(&mut self, factor: SimdReal)

Dilate all the Aabbs represented by self`` by their extents multiplied by the given scale factor`.

pub fn replace(&mut self, i: usize, aabb: Aabb)

Replace the i-th Aabb of this SIMD AAAB by the given value.

pub fn cast_local_ray( &self, ray: &SimdRay, max_toi: SimdReal ) -> (SimdBool, SimdReal)

Casts a ray on all the Aabbs represented by self.

pub fn distance_to_local_point(&self, point: &Point<SimdReal>) -> SimdReal

Computes the distances between a point and all the Aabbs represented by self.

pub fn distance_to_origin(&self) -> SimdReal

Computes the distances between the origin and all the Aabbs represented by self.

pub fn contains_local_point(&self, point: &Point<SimdReal>) -> SimdBool

Check which Aabb represented by self contains the given point.

pub fn contains(&self, other: &SimdAabb) -> SimdBool

Lanewise check which Aabb represented by self contains the given set of other aabbs. The check is performed lane-wise.

pub fn intersects(&self, other: &SimdAabb) -> SimdBool

Check which Aabb represented by self contains the given set of other aabbs. The check is performed lane-wise.

pub fn intersects_permutations(&self, other: &SimdAabb) -> [SimdBool; 4]

Checks intersections between all the lanes combination between self and other.

The result is an array such that result[i].extract(j) contains the intersection result between self.extract(i) and other.extract(j).

pub fn to_merged_aabb(&self) -> Aabb

Merge all the Aabb represented by self into a single one.

pub fn extract(&self, lane: usize) -> Aabb

Extracts the Aabb stored in the given SIMD lane of the SIMD Aabb: