Struct parry3d::mass_properties::MassProperties

pub struct MassProperties {
    pub local_com: Point<Real>,
    pub inv_mass: Real,
    pub inv_principal_inertia_sqrt: AngVector<Real>,
    pub principal_inertia_local_frame: Rotation<Real>,
}

The local mass properties of a rigid-body.

Fields
local_com: Point<Real>

The center of mass of a rigid-body expressed in its local-space.

inv_mass: Real

The inverse of the mass of a rigid-body.

If this is zero, the rigid-body is assumed to have infinite mass.

inv_principal_inertia_sqrt: AngVector<Real>

The inverse of the principal angular inertia of the rigid-body.

Components set to zero are assumed to be infinite along the corresponding principal axis.

principal_inertia_local_frame: Rotation<Real>

The principal vectors of the local angular inertia tensor of the rigid-body.

Implementations

impl MassProperties

pub fn new(
local_com: Point<Real>,
mass: Real,
principal_inertia: AngVector<Real>
) -> Self

Initializes the mass properties from the given center-of-mass, mass, and principal angular inertia.

The center-of-mass is specified in the local-space of the rigid-body. The principal angular inertia are the angular inertia along the coordinate axes in the local-space of the rigid-body.

pub fn with_principal_inertia_frame(
local_com: Point<Real>,
mass: Real,
principal_inertia: AngVector<Real>,
principal_inertia_local_frame: Rotation<Real>
) -> Self

Initializes the mass properties from the given center-of-mass, mass, and principal angular inertia.

The center-of-mass is specified in the local-space of the rigid-body. The principal angular inertia are the angular inertia along the coordinate axes defined by the principal_inertia_local_frame expressed in the local-space of the rigid-body.

pub fn with_inertia_matrix(
local_com: Point<Real>,
mass: Real,
inertia: Matrix3<Real>
) -> Self

Initialize a new MassProperties from a given center-of-mass, mass, and angular inertia matrix.

The angular inertia matrix will be diagonalized in order to extract the principal inertia values and principal inertia frame.

pub fn mass(&self) -> Real

The mass.

pub fn principal_inertia(&self) -> AngVector<Real>

The angular inertia along the principal inertia axes of the rigid-body.

pub fn world_com(&self, pos: &Isometry<Real>) -> Point<Real>

The world-space center of mass of the rigid-body.

pub fn world_inv_inertia_sqrt(
&self,
rot: &Rotation<Real>
) -> AngularInertia<Real>

The world-space inverse angular inertia tensor of the rigid-body.

pub fn reconstruct_inverse_inertia_matrix(&self) -> Matrix3<Real>

Reconstructs the inverse angular inertia tensor of the rigid body from its principal inertia values and axes.

pub fn reconstruct_inertia_matrix(&self) -> Matrix3<Real>

Reconstructs the angular inertia tensor of the rigid body from its principal inertia values and axes.

pub fn transform_by(&self, m: &Isometry<Real>) -> Self

Transform each element of the mass properties.

pub fn set_mass(&mut self, new_mass: Real, adjust_angular_inertia: bool)

Changes the mass on these mass-properties.

The adjust_angular_inertia argument should always be true, unless there are some specific reasons not to do so. Setting this to true will automatically adjust the angular inertia of self to account for the mass change (i.e. it will multiply the angular inertia by new_mass / prev_mass). Setting it to false will not change the current angular inertia.

impl MassProperties

pub fn from_ball(density: Real, radius: Real) -> Self

Computes the mass properties of a ball.

impl MassProperties

pub fn from_capsule(
density: Real,
a: Point<Real>,
b: Point<Real>,
radius: Real
) -> Self

Computes the mass properties of a capsule.

impl MassProperties

pub fn from_compound(
density: Real,
shapes: &[(Isometry<Real>, SharedShape)]
) -> Self

Computes the mass properties of a compound shape.

impl MassProperties

pub fn from_cone(density: Real, half_height: Real, radius: Real) -> Self

Computes the mass properties of a cone.

impl MassProperties

pub fn from_convex_polyhedron(
density: Real,
vertices: &[Point<Real>],
indices: &[[u32; 3]]
) -> MassProperties

Computes the mass properties of a convex polyhedron.

impl MassProperties

pub fn from_cuboid(density: Real, half_extents: Vector<Real>) -> Self

Computes the mass properties of a cuboid.

impl MassProperties

pub fn from_cylinder(density: Real, half_height: Real, radius: Real) -> Self

Computes the mass properties of a cylinder.

impl MassProperties

pub fn from_trimesh(
density: Real,
vertices: &[Point<Real>],
indices: &[[u32; 3]]
) -> MassProperties

Computes the mass properties of a triangle mesh.