Type Alias parry3d::math::SdpMatrix

pub type SdpMatrix<N> = SdpMatrix3<N>;

A 3D symmetric-definite-positive matrix.

Aliased Type

struct SdpMatrix<N> {
    pub m11: N,
    pub m12: N,
    pub m13: N,
    pub m22: N,
    pub m23: N,
    pub m33: N,
}

Fields
m11: N

The component at the first row and first column of this matrix.

m12: N

The component at the first row and second column of this matrix.

m13: N

The component at the first row and third column of this matrix.

m22: N

The component at the second row and second column of this matrix.

m23: N

The component at the second row and third column of this matrix.

m33: N

The component at the third row and third column of this matrix.

Implementations

impl<N: SimdRealField + Copy> SdpMatrix3<N>

pub fn new(m11: N, m12: N, m13: N, m22: N, m23: N, m33: N) -> Self

A new SDP 3x3 matrix with the given components.

Because the matrix is symmetric, only the lower off-diagonal components is required.

pub fn from_sdp_matrix(mat: Matrix3<N>) -> Self

Build an SdpMatrix3 structure from a plain matrix, assuming it is SDP.

No check is performed to ensure mat is actually SDP.

pub fn zero() -> Self

Create a new SDP matrix filled with zeros.

pub fn diagonal(val: N) -> Self

Create a new SDP matrix with its diagonal filled with val, and its off-diagonal elements set to zero.

pub fn is_zero(&self) -> bool

Are all components of this matrix equal to zero?

pub fn inverse_unchecked(&self) -> Self

Compute the inverse of this SDP matrix without performing any inversibility check.

pub fn quadform3x2(&self, m: &Matrix3x2<N>) -> SdpMatrix2<N>

Compute the quadratic form m.transpose() * self * m.

pub fn quadform(&self, m: &Matrix3<N>) -> Self

Compute the quadratic form m.transpose() * self * m.

pub fn add_diagonal(&self, elt: N) -> Self

Adds elt to the diagonal components of self.