Module parry3d::partitioning

Spatial partitioning tools.

Structs

QBVH

A quaternary bounding-volume-hierarchy.

QbvhNonOverlappingDataSplitter

Data splitter for QBVH construction that generates non-overlapping AABBs at each level of the tree.

Enums

SimdBestFirstVisitStatus

The next action to be taken by a BVH traversal algorithm after having visited a node with some data.

SimdSimultaneousVisitStatus

The status of the simultaneous traversal of two spatial partitioning structures.

SimdVisitStatus

The status of the spatial partitioning structure traversal.

Traits

IndexedData

A data to which an index is associated.

QBVHDataGenerator

Trait used for generating the content of the leaves of the QBVH acceleration structure.

SimdBestFirstVisitor

Trait implemented by cost functions used by the best-first search on a BVT.

SimdSimultaneousVisitor

Trait implemented by visitor called during a simultaneous spatial partitioning data structure tarversal.

SimdVisitor

Trait implemented by visitor called during the traversal of a spatial partitioning data structure.

Type Definitions

SimdQbvhDeprecated

A quaternary bounding-volume-hierarchy.