Struct parry3d::partitioning::QBVH

pub struct QBVH<T> { /* private fields */ }

A quaternary bounding-volume-hierarchy.

This is a bounding-volume-hierarchy where each node has either four children or none.

Implementations

impl<T: IndexedData> QBVH<T>

pub fn clear_and_rebuild(
&mut self,
data_gen: impl QBVHDataGenerator<T>,
dilation_factor: Real
)

Clears this quaternary BVH and rebuilds it from a new set of data and AABBs.

impl<T: IndexedData> QBVH<T>

pub fn clear_and_rebuild_with_splitter(
&mut self,
data_gen: impl QBVHDataGenerator<T>,
splitter: impl QBVHDataSplitter<T>,
dilation_factor: Real
)

Clears this quaternary BVH and rebuilds it from a new set of data and AABBs.

impl<T: IndexedData> QBVH<T>

pub fn new() -> Self

Initialize an empty QBVH.

pub fn iter_data_mut(&mut self) -> impl Iterator<Item = (NodeIndex, &mut T)>

Iterates mutably through all the leaf data in this QBVH.

pub fn iter_data(&self) -> impl Iterator<Item = (NodeIndex, &T)>

Iterate through all the leaf data in this QBVH.

pub fn root_aabb(&self) -> &AABB

The AABB of the root of this tree.

pub fn node_aabb(&self, node_id: NodeIndex) -> Option<AABB>

The AABB of the given node.

pub fn leaf_data(&mut self, node_id: NodeIndex) -> Option<T>

Returns the data associated to a given leaf.

Returns None if the provided node ID does not identify a leaf.

pub fn raw_nodes(&self) -> &[QBVHNode]

The raw nodes of this BVH.

If this QBVH isn’t empty, the first element of the returned slice is the root of the tree. The other elements are not arranged in any particular order. The more high-level traversal methods should be used instead of this.

pub fn raw_proxies(&self) -> &[QBVHProxy<T>]

The raw proxies of this BVH.

If this QBVH isn’t empty, the first element of the returned slice is the root of the tree. The other elements are not arranged in any particular order. The more high-level traversal methods should be used instead of this.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Computes a scaled version of this QBVH.

This will apply the scale to each AABB on this BVH.

impl<T: IndexedData> QBVH<T>

pub fn traverse_depth_first(&self, visitor: &mut impl SimdVisitor<T, SimdAABB>)

Performs a depth-first traversal on the BVH.

pub fn traverse_depth_first_with_stack(
&self,
visitor: &mut impl SimdVisitor<T, SimdAABB>,
stack: &mut Vec<u32>
)

Performs a depth-first traversal on the BVH.

pub fn traverse_best_first<BFS>(
&self,
visitor: &mut BFS
) -> Option<(NodeIndex, BFS::Result)> where
BFS: SimdBestFirstVisitor<T, SimdAABB>,
BFS::Result: Clone,

Performs a best-first-search on the BVH.

Returns the content of the leaf with the smallest associated cost, and a result of user-defined type.

pub fn intersect_aabb(&self, aabb: &AABB, out: &mut Vec<T>)

Retrieve all the data of the nodes with AABBs intersecting the given AABB:

pub fn traverse_bvtt<T2: IndexedData>(
&self,
qbvh2: &QBVH<T2>,
visitor: &mut impl SimdSimultaneousVisitor<T, T2, SimdAABB>
)

Performs a simultaneous traversal of two QBVH.

pub fn traverse_bvtt_with_stack<T2: IndexedData>(
&self,
qbvh2: &QBVH<T2>,
visitor: &mut impl SimdSimultaneousVisitor<T, T2, SimdAABB>,
stack: &mut Vec<(u32, u32)>
)

Performs a simultaneous traversal of two QBVH.

impl<T: IndexedData> QBVH<T>

pub fn pre_update(&mut self, data: T)

Marks a piece of data as dirty so it can be updated during the next call to self.update.

pub fn update<F>(&mut self, aabb_builder: F, dilation_factor: Real) where
F: Fn(&T) -> AABB,

Update all the nodes that have been marked as dirty by self.pre_update.