Trait parry3d::partitioning::SimdSimultaneousVisitor

pub trait SimdSimultaneousVisitor<T1, T2, SimdBV> {
    fn visit(
        &mut self, 
        left_bv: &SimdBV, 
        left_data: Option<[Option<&T1>; 4]>, 
        right_bv: &SimdBV, 
        right_data: Option<[Option<&T2>; 4]>
    ) -> SimdSimultaneousVisitStatus;
}

Trait implemented by visitor called during a simultaneous spatial partitioning data structure tarversal.

Required Methods

fn visit(
&mut self,
left_bv: &SimdBV,
left_data: Option<[Option<&T1>; 4]>,
right_bv: &SimdBV,
right_data: Option<[Option<&T2>; 4]>
) -> SimdSimultaneousVisitStatus

Execute an operation on the content of two nodes, one from each structure.

Returns whether the traversal should continue on the nodes children, if it should not continue on those children, or if the whole traversal should be exited early.