Trait parry3d::partitioning::SimdVisitor

pub trait SimdVisitor<LeafData, SimdBV> {
    // Required method
    fn visit(
        &mut self,
        bv: &SimdBV,
        data: Option<[Option<&LeafData>; 4]>
    ) -> SimdVisitStatus;
}

Trait implemented by visitor called during the traversal of a spatial partitioning data structure.

Required Methods

fn visit( &mut self, bv: &SimdBV, data: Option<[Option<&LeafData>; 4]> ) -> SimdVisitStatus

Execute an operation on the content of a node of the spatial partitioning structure.

Returns whether the traversal should continue on the node’s children, if it should not continue on those children, or if the whole traversal should be exited early.