Type Definition parry3d::partitioning::Qbvh

pub type Qbvh<LeafData> = GenericQbvh<LeafData, DefaultStorage>;

A quaternary bounding-volume-hierarchy.

This is a bounding-volume-hierarchy where each node has either four children or none.

Implementations

impl<LeafData: IndexedData> Qbvh<LeafData>

pub fn new() -> Self

Initialize an empty Qbvh.

pub fn iter_data_mut( &mut self ) -> impl Iterator<Item = (NodeIndex, &mut LeafData)>

Iterates mutably through all the leaf data in this Qbvh.

pub fn iter_data(&self) -> impl Iterator<Item = (NodeIndex, &LeafData)>

Iterate through all the leaf data in this Qbvh.

pub fn node_aabb(&self, node_id: NodeIndex) -> Option<Aabb>

The Aabb of the given node.

pub fn leaf_data(&self, node_id: NodeIndex) -> Option<LeafData>

Returns the data associated to a given leaf.

Returns None if the provided node ID does not identify a leaf.

pub fn raw_nodes(&self) -> &[QbvhNode]

The raw nodes of this BVH.

If this Qbvh isn’t empty, the first element of the returned slice is the root of the tree. The other elements are not arranged in any particular order. The more high-level traversal methods should be used instead of this.

pub fn raw_proxies(&self) -> &[QbvhProxy<LeafData>]

The raw proxies of this BVH.

If this Qbvh isn’t empty, the first element of the returned slice is the root of the tree. The other elements are not arranged in any particular order. The more high-level traversal methods should be used instead of this.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Computes a scaled version of this Qbvh.

This will apply the scale to each Aabb on this BVH.

impl<LeafData: IndexedData> Qbvh<LeafData>

pub fn clear_and_rebuild( &mut self, data_gen: impl QbvhDataGenerator<LeafData>, dilation_factor: Real )

Clears this quaternary BVH and rebuilds it from a new set of data and Aabbs.

impl<LeafData: IndexedData> Qbvh<LeafData>

pub fn clear_and_rebuild_with_splitter( &mut self, data_gen: impl QbvhDataGenerator<LeafData>, splitter: impl QbvhDataSplitter<LeafData>, dilation_factor: Real )

Clears this quaternary BVH and rebuilds it from a new set of data and Aabbs.

impl<LeafData: IndexedData> Qbvh<LeafData>

pub fn remove(&mut self, data: LeafData) -> Option<LeafData>

Immediately remove a leaf from this QBVH.

pub fn pre_update_or_insert(&mut self, data: LeafData)

Prepare a new leaf for insertion into this QBVH (or for update if it already exists).

The insertion or update will be completely valid only after the next call to Qbvh::refit.

pub fn refit<F>( &mut self, margin: Real, workspace: &mut QbvhUpdateWorkspace, aabb_builder: F ) -> usizewhere F: Fn(&LeafData) -> Aabb,

Update all the nodes that have been marked as dirty by Qbvh::pre_update_or_insert, and Qbvh::remove.

This will not alter the topology of this Qbvh.

pub fn rebalance(&mut self, margin: Real, workspace: &mut QbvhUpdateWorkspace)

Rebalances the Qbvh tree.

This will modify the topology of this tree. This assumes that the leaf AABBs have already been updated with Qbvh::refit.