Module parry3d::query::closest_points

Implementation details of the closest_points function.

Structs

CompositeShapeAgainstShapeClosestPointsVisitor

A visitor for computing the closest points between a composite-shape and a shape.

Enums

ClosestPoints

Closest points information.

Functions

closest_points

Computes the pair of closest points between two shapes.

closest_points_ball_ball

Closest points between balls.

closest_points_ball_convex_polyhedron

ClosestPoints between a ball and a convex polyhedron.

closest_points_composite_shape_shape

Closest points between a composite shape and any other shape.

closest_points_convex_polyhedron_ball

ClosestPoints between a convex polyhedron and a ball.

closest_points_cuboid_cuboid

Closest points between two cuboids.

closest_points_cuboid_triangle

Closest points between a cuboid and a triangle.

closest_points_halfspace_support_map

Closest points between a halfspace and a support-mapped shape (Cuboid, ConvexHull, etc.)

closest_points_line_line

Closest points between two segments.

closest_points_line_line_parameters

Closest points between two lines.

closest_points_line_line_parameters_eps

Closest points between two lines with a custom tolerance epsilon.

closest_points_segment_segment

Closest points between segments.

closest_points_segment_segment_with_locations

Closest points between two segments.

closest_points_segment_segment_with_locations_nD

Segment-segment closest points computation in an arbitrary dimension.

closest_points_shape_composite_shape

Closest points between a shape and a composite shape.

closest_points_support_map_halfspace

Closest points between a support-mapped shape (Cuboid, ConvexHull, etc.) and a halfspace.

closest_points_support_map_support_map

Closest points between support-mapped shapes (Cuboid, ConvexHull, etc.)

closest_points_support_map_support_map_with_params

Closest points between support-mapped shapes (Cuboid, ConvexHull, etc.)

closest_points_triangle_cuboid

Closest points between a triangle and a cuboid.