Struct parry3d::query::details::TOI

pub struct TOI {
    pub toi: Real,
    pub witness1: Point<Real>,
    pub witness2: Point<Real>,
    pub normal1: Unit<Vector<Real>>,
    pub normal2: Unit<Vector<Real>>,
    pub status: TOIStatus,
}

The result of a time-of-impact (TOI) computation.

Fields
toi: Real

The time at which the objects touch.

witness1: Point<Real>

The local-space closest point on the first shape at the time of impact.

Undefined if status is Penetrating.

witness2: Point<Real>

The local-space closest point on the second shape at the time of impact.

Undefined if status is Penetrating.

normal1: Unit<Vector<Real>>

The local-space outward normal on the first shape at the time of impact.

Undefined if status is Penetrating.

normal2: Unit<Vector<Real>>

The local-space outward normal on the second shape at the time of impact.

Undefined if status is Penetrating.

status: TOIStatus

The way the time-of-impact computation algorithm terminated.

Implementations

impl TOI

pub fn swapped(self) -> Self

Swaps every data of this TOI result such that the role of both shapes are inverted.

In practice, this makes it so that self.witness1 and self.normal1 become self.witness2 and self.normal2 and vice-versa.

pub fn transform1_by(&self, pos: &Isometry<Real>) -> Self

Transform self.witness1 and self.normal1 by pos.