Trait parry3d::query::details::RayCast

pub trait RayCast {
    fn cast_local_ray_and_get_normal(
        &self, 
        ray: &Ray, 
        max_toi: Real, 
        solid: bool
    ) -> Option<RayIntersection>;

    fn cast_local_ray(
        &self, 
        ray: &Ray, 
        max_toi: Real, 
        solid: bool
    ) -> Option<Real> { ... }
    fn intersects_local_ray(&self, ray: &Ray, max_toi: Real) -> bool { ... }
    fn cast_ray(
        &self, 
        m: &Isometry<Real>, 
        ray: &Ray, 
        max_toi: Real, 
        solid: bool
    ) -> Option<Real> { ... }
    fn cast_ray_and_get_normal(
        &self, 
        m: &Isometry<Real>, 
        ray: &Ray, 
        max_toi: Real, 
        solid: bool
    ) -> Option<RayIntersection> { ... }
    fn intersects_ray(
        &self, 
        m: &Isometry<Real>, 
        ray: &Ray, 
        max_toi: Real
    ) -> bool { ... }
}

Traits of objects which can be transformed and tested for intersection with a ray.

Required Methods

fn cast_local_ray_and_get_normal(
&self,
ray: &Ray,
max_toi: Real,
solid: bool
) -> Option<RayIntersection>

Computes the time of impact, and normal between this transformed shape and a ray.

Provided Methods

fn cast_local_ray(&self, ray: &Ray, max_toi: Real, solid: bool) -> Option<Real>

Computes the time of impact between this transform shape and a ray.

fn intersects_local_ray(&self, ray: &Ray, max_toi: Real) -> bool

Tests whether a ray intersects this transformed shape.

fn cast_ray(
&self,
m: &Isometry<Real>,
ray: &Ray,
max_toi: Real,
solid: bool
) -> Option<Real>

Computes the time of impact between this transform shape and a ray.

fn cast_ray_and_get_normal(
&self,
m: &Isometry<Real>,
ray: &Ray,
max_toi: Real,
solid: bool
) -> Option<RayIntersection>

Computes the time of impact, and normal between this transformed shape and a ray.

fn intersects_ray(&self, m: &Isometry<Real>, ray: &Ray, max_toi: Real) -> bool

Tests whether a ray intersects this transformed shape.