Function parry3d::query::gjk::project_origin

pub fn project_origin<G: ?Sized>(
    m: &Isometry<Real>, 
    g: &G, 
    simplex: &mut VoronoiSimplex
) -> Option<Point<Real>> where
    G: SupportMap, 

Projects the origin on the boundary of the given shape.

The origin is assumed to be outside of the shape. If it is inside, use the EPA algorithm instead. Return None if the origin is not inside of the shape or if the EPA algorithm failed to compute the projection.

Return the projected point in the local-space of g.