Struct parry3d::query::gjk::CSOPoint

pub struct CSOPoint {
    pub point: Point<Real>,
    pub orig1: Point<Real>,
    pub orig2: Point<Real>,
}

A point of a Configuration-Space Obstacle.

A Configuration-Space Obstacle (CSO) is the result of the Minkowski Difference of two solids. In other words, each of its points correspond to the difference of two point, each belonging to a different solid.

Fields
point: Point<Real>

The point on the CSO. This is equal to self.orig1 - self.orig2, unless this CSOPoint has been translated with self.translate.

orig1: Point<Real>

The original point on the first shape used to compute self.point.

orig2: Point<Real>

The original point on the second shape used to compute self.point.

Implementations

impl CSOPoint

pub fn new(orig1: Point<Real>, orig2: Point<Real>) -> Self

Initializes a CSO point with orig1 - orig2.

pub fn new_with_point( point: Point<Real>, orig1: Point<Real>, orig2: Point<Real> ) -> Self

Initializes a CSO point with all information provided.

It is assumed, but not checked, that point == orig1 - orig2.

pub fn single_point(point: Point<Real>) -> Self

Initializes a CSO point where both original points are equal.

pub fn origin() -> Self

CSO point where all components are set to zero.

pub fn from_shapes_toward<G1, G2>( pos12: &Isometry<Real>, g1: &G1, g2: &G2, dir: &Unit<Vector<Real>> ) -> Selfwhere G1: SupportMap + ?Sized, G2: SupportMap + ?Sized,

Computes the support point of the CSO of g1 and g2 toward the unit direction dir.

pub fn from_shapes<G1, G2>( pos12: &Isometry<Real>, g1: &G1, g2: &G2, dir: &Vector<Real> ) -> Selfwhere G1: SupportMap + ?Sized, G2: SupportMap + ?Sized,

Computes the support point of the CSO of g1 and g2 toward the direction dir.

pub fn translate(&self, dir: &Vector<Real>) -> Self

Translate the CSO point.

pub fn translate_mut(&mut self, dir: &Vector<Real>)

Translate in-place the CSO point.