Trait parry3d::query::point::PointQuery

pub trait PointQuery {
    // Required methods
    fn project_local_point(
        &self,
        pt: &Point<Real>,
        solid: bool
    ) -> PointProjection;
    fn project_local_point_and_get_feature(
        &self,
        pt: &Point<Real>
    ) -> (PointProjection, FeatureId);

    // Provided methods
    fn project_local_point_with_max_dist(
        &self,
        pt: &Point<Real>,
        solid: bool,
        max_dist: Real
    ) -> Option<PointProjection> { ... }
    fn project_point_with_max_dist(
        &self,
        m: &Isometry<Real>,
        pt: &Point<Real>,
        solid: bool,
        max_dist: Real
    ) -> Option<PointProjection> { ... }
    fn distance_to_local_point(&self, pt: &Point<Real>, solid: bool) -> Real { ... }
    fn contains_local_point(&self, pt: &Point<Real>) -> bool { ... }
    fn project_point(
        &self,
        m: &Isometry<Real>,
        pt: &Point<Real>,
        solid: bool
    ) -> PointProjection { ... }
    fn distance_to_point(
        &self,
        m: &Isometry<Real>,
        pt: &Point<Real>,
        solid: bool
    ) -> Real { ... }
    fn project_point_and_get_feature(
        &self,
        m: &Isometry<Real>,
        pt: &Point<Real>
    ) -> (PointProjection, FeatureId) { ... }
    fn contains_point(&self, m: &Isometry<Real>, pt: &Point<Real>) -> bool { ... }
}

Trait of objects that can be tested for point inclusion and projection.

Required Methods

fn project_local_point(&self, pt: &Point<Real>, solid: bool) -> PointProjection

Projects a point on self.

The point is assumed to be expressed in the local-space of self.

fn project_local_point_and_get_feature( &self, pt: &Point<Real> ) -> (PointProjection, FeatureId)

Projects a point on the boundary of self and returns the id of the feature the point was projected on.

Provided Methods

fn project_local_point_with_max_dist( &self, pt: &Point<Real>, solid: bool, max_dist: Real ) -> Option<PointProjection>

Projects a point on self, unless the projection lies further than the given max distance.

The point is assumed to be expressed in the local-space of self.

fn project_point_with_max_dist( &self, m: &Isometry<Real>, pt: &Point<Real>, solid: bool, max_dist: Real ) -> Option<PointProjection>

Projects a point on self transformed by m, unless the projection lies further than the given max distance.

fn distance_to_local_point(&self, pt: &Point<Real>, solid: bool) -> Real

Computes the minimal distance between a point and self.

fn contains_local_point(&self, pt: &Point<Real>) -> bool

Tests if the given point is inside of self.

fn project_point( &self, m: &Isometry<Real>, pt: &Point<Real>, solid: bool ) -> PointProjection

Projects a point on self transformed by m.

fn distance_to_point( &self, m: &Isometry<Real>, pt: &Point<Real>, solid: bool ) -> Real

Computes the minimal distance between a point and self transformed by m.

fn project_point_and_get_feature( &self, m: &Isometry<Real>, pt: &Point<Real> ) -> (PointProjection, FeatureId)

Projects a point on the boundary of self transformed by m and returns the id of the feature the point was projected on.

fn contains_point(&self, m: &Isometry<Real>, pt: &Point<Real>) -> bool

Tests if the given point is inside of self transformed by m.