Module parry3d::query::sat

Application of the Separating-Axis-Theorem (SAT).

Functions

cuboid_cuboid_compute_separation_wrt_local_line

Computes the separation of two cuboids along axis1.

cuboid_cuboid_find_local_separating_edge_twoway

Finds the best separating edge between two cuboids.

cuboid_cuboid_find_local_separating_normal_oneway

Finds the best separating normal between two cuboids.

cuboid_segment_find_local_separating_edge_twoway

Finds the best separating edge between a cuboid and a segment.

cuboid_support_map_compute_separation_wrt_local_line

Computes the separation between a cuboid an a convex shape implementing the SupportMap trait, along the given axis.

cuboid_support_map_find_local_separating_edge_twoway

Finds the best separating edge between a cuboid and a convex shape implementing the Supportmap trait.

cuboid_support_map_find_local_separating_normal_oneway

Finds the best separating normal between a cuboid and a convex shape implementing the SupportMap trait.

cuboid_triangle_find_local_separating_edge_twoway

Finds the best separating edge between a cuboid and a triangle.

point_cuboid_find_local_separating_normal_oneway

Computes the separation between a point and a cuboid, along the given direction normal1.

segment_triangle_find_local_separating_edge_twoway

Finds the best separating edge between a segment and a triangle.

support_map_support_map_compute_separation

Computes the separation along the given direction, between two convex shapes implementing the SupportMap trait.

triangle_cuboid_find_local_separating_normal_oneway

Finds the best separating normal a triangle and a cuboid.

triangle_segment_find_local_separating_normal_oneway

Finds the best separating normal a triangle and a segment.