Struct parry3d::query::TrackedContact

pub struct TrackedContact<Data> {
    pub local_p1: Point<Real>,
    pub local_p2: Point<Real>,
    pub dist: Real,
    pub fid1: u32,
    pub fid2: u32,
    pub data: Data,
}

A single contact between two shape.

Fields
local_p1: Point<Real>

The contact point in the local-space of the first shape.

local_p2: Point<Real>

The contact point in the local-space of the second shape.

dist: Real

The distance between the two contact points.

fid1: u32

The feature ID of the first shape involved in the contact.

fid2: u32

The feature ID of the second shape involved in the contact.

data: Data

User-data associated to this contact.

Implementations

impl<Data: Default + Copy> TrackedContact<Data>

pub fn new(
local_p1: Point<Real>,
local_p2: Point<Real>,
fid1: u32,
fid2: u32,
dist: Real
) -> Self

Creates a new tracked contact.

pub fn flipped(
local_p1: Point<Real>,
local_p2: Point<Real>,
fid1: u32,
fid2: u32,
dist: Real,
flipped: bool
) -> Self

Creates a new tracked contact where its input may need to be flipped.

pub fn copy_geometry_from(&mut self, contact: Self)

Copy to self the geometric information from contact.