Module parry3d::shape

Shapes supported by parry.

Structs

Ball

A Ball shape.

Capsule

A capsule shape defined as a round segment.

Compound

A compound shape with an aabb bounding volume.

Cone

Cone shape with its principal axis aligned with the y axis.

ConvexPolyhedron

A convex polyhedron without degenerate faces.

Cuboid

Shape of a box.

Cylinder

Cylinder shape with its principal axis aligned with the y axis.

GenericHeightField

A 3D heightfield with a generic storage buffer for its height grid.

HalfSpace

A half-space delimited by an infinite plane.

HeightFieldCellStatus

The status of the cell of an heightfield.

HeightFieldRadialTriangles

An iterator through all the triangles around the given point, after vertical projection on the heightfield.

PolygonalFeature

A polygonal feature representing the local polygonal approximation of a vertex, face, or edge of a convex shape.

Polyline

A polyline.

RoundShape

A shape with rounded borders.

Segment

A segment shape.

SharedShape

The shape of a collider.

Tetrahedron

A tetrahedron with 4 vertices.

TopoFace

A face of a triangle-mesh’s half-edge topology.

TopoHalfEdge

A half-edge of a triangle-mesh’s half-edge topology.

TopoVertex

A vertex of a triangle-mesh’s half-edge topology.

TriMesh

A triangle mesh.

TriMeshConnectedComponents

The connected-components of a triangle mesh.

TriMeshFlags

The status of the cell of an heightfield.

TriMeshPseudoNormals

The set of pseudo-normals of a triangle mesh.

TriMeshTopology

The half-edge topology information of a triangle mesh.

Triangle

A triangle shape.

Enums

FeatureId

An identifier of a feature of a convex polyhedron.

SegmentPointLocation

Logical description of the location of a point on a triangle.

ShapeType

Enum representing the type of a shape.

TetrahedronPointLocation

Logical description of the location of a point on a triangle.

TopologyError

Indicated an inconsistency in the topology of a triangle mesh.

TriangleOrientation

Orientation of a triangle.

TrianglePointLocation

Description of the location of a point on a triangle.

TypedShape

Enum representing the shape with its actual type

Traits

HeightFieldStorage

Abstraction over the storage type of an heightfield’s heights grid.

PolygonalFeatureMap

Trait implemented by convex shapes with features with polyhedral approximations.

Shape

Trait implemented by shapes usable by Rapier.

SimdCompositeShape

Trait implemented by shapes composed of multiple simpler shapes.

SupportMap

Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Definitions

HeightField

A 3D heightfield.

RoundCone

A cone dilated by a sphere (so it has round corners).

RoundConvexPolyhedron

A convex polyhedron dilated by a sphere (so it has round corners).

RoundCuboid

A cuboid dilated by a sphere (so it has round corners).

RoundCylinder

A cylinder dilated by a sphere (so it has round corners).

RoundTriangle

A triangle dilated by a sphere (so it has round corners).