Struct parry3d::shape::Ball

#[repr(C)]
pub struct Ball {
    pub radius: Real,
}

A Ball shape.

Fields
radius: Real

The radius of the ball.

Implementations

impl Ball

pub fn aabb(&self, pos: &Isometry<Real>) -> AABB

Computes the world-space AABB of this ball transformed by pos.

pub fn local_aabb(&self) -> AABB

Computes the local-space AABB of this ball.

impl Ball

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this ball, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space AABB of this ball.

impl Ball

pub fn new(radius: Real) -> Ball

Creates a new ball from its radius and center.

pub fn scaled(
self,
scale: &Vector<Real>,
nsubdivs: u32
) -> Option<Either<Self, ConvexPolyhedron>>

Computes a scaled version of this ball.

If the scaling factor is non-uniform, then it can’t be represented as ball. Instead, a convex polygon approximation (with nsubdivs subdivisions) is returned. Returns None if that approximation had degenerate normals (for example if the scaling factor along one axis is zero).

impl Ball

pub fn to_outline(&self, nsubdiv: u32) -> (Vec<Point3<Real>>, Vec<[u32; 2]>)

Outlines this ball’s shape using polylines.

impl Ball

pub fn to_trimesh(
&self,
ntheta_subdiv: u32,
nphi_subdiv: u32
) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this ball as a triangle-mesh.