Struct parry3d::shape::Capsule

#[repr(C)]
pub struct Capsule {
    pub segment: Segment,
    pub radius: Real,
}

A capsule shape defined as a round segment.

Fields
segment: Segment

The endpoints of the capsule’s principal axis.

radius: Real

The radius of the capsule.

Implementations

impl Capsule

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

The axis-aligned bounding box of this capsule.

pub fn local_aabb(&self) -> Aabb

The axis-aligned bounding box of this capsule.

impl Capsule

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this capsule, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the world-space bounding sphere of this capsule.

impl Capsule

pub fn new_x(half_height: Real, radius: Real) -> Self

Creates a new capsule aligned with the x axis and with the given half-height an radius.

pub fn new_y(half_height: Real, radius: Real) -> Self

Creates a new capsule aligned with the y axis and with the given half-height an radius.

pub fn new_z(half_height: Real, radius: Real) -> Self

Creates a new capsule aligned with the z axis and with the given half-height an radius.

pub fn new(a: Point<Real>, b: Point<Real>, radius: Real) -> Self

Creates a new capsule defined as the segment between a and b and with the given radius.

pub fn height(&self) -> Real

The height of this capsule.

pub fn half_height(&self) -> Real

The half-height of this capsule.

pub fn center(&self) -> Point<Real>

The center of this capsule.

pub fn transform_by(&self, pos: &Isometry<Real>) -> Self

Creates a new capsule equal to self with all its endpoints transformed by pos.

pub fn canonical_transform(&self) -> Isometry<Real>

The transformation such that t * Y is collinear with b - a and t * origin equals the capsule’s center.

pub fn rotation_wrt_y(&self) -> Rotation<Real>

The rotation r such that r * Y is collinear with b - a.

pub fn transform_wrt_y(&self) -> Isometry<Real>

The transform t such that t * Y is collinear with b - a and such that t * origin = (b + a) / 2.0.

pub fn scaled( self, scale: &Vector<Real>, nsubdivs: u32 ) -> Option<Either<Self, ConvexPolyhedron>>

Computes a scaled version of this capsule.

If the scaling factor is non-uniform, then it can’t be represented as capsule. Instead, a convex polygon approximation (with nsubdivs subdivisions) is returned. Returns None if that approximation had degenerate normals (for example if the scaling factor along one axis is zero).

impl Capsule

pub fn to_outline(&self, nsubdiv: u32) -> (Vec<Point3<Real>>, Vec<[u32; 2]>)

Outlines this capsule’s shape using polylines.

impl Capsule

pub fn to_trimesh( &self, ntheta_subdiv: u32, nphi_subdiv: u32 ) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this capsule as a triangle-mesh.