Struct parry3d::shape::Compound

pub struct Compound { /* private fields */ }

A compound shape with an aabb bounding volume.

A compound shape is a shape composed of the union of several simpler shape. This is the main way of creating a concave shape from convex parts. Each parts can have its own delta transformation to shift or rotate it with regard to the other shapes.

Implementations

impl Compound

pub fn new(shapes: Vec<(Isometry<Real>, SharedShape)>) -> Compound

Builds a new compound shape.

Panics if the input vector is empty, of if some of the provided shapes are also composite shapes (nested composite shapes are not allowed).

impl Compound

pub fn shapes(&self) -> &[(Isometry<Real>, SharedShape)]

The shapes of this compound shape.

pub fn local_aabb(&self) -> &Aabb

The Aabb of this compound in its local-space.

pub fn local_bounding_sphere(&self) -> BoundingSphere

The bounding-sphere of this compound in its local-space.

pub fn aabbs(&self) -> &[Aabb]

The shapes Aabbs.

pub fn qbvh(&self) -> &Qbvh<u32>

The acceleration structure used by this compound shape.