Struct parry3d::shape::ConvexPolyhedron

pub struct ConvexPolyhedron { /* private fields */ }

A convex polyhedron without degenerate faces.

Implementations

impl ConvexPolyhedron

pub fn aabb(&self, pos: &Isometry<Real>) -> AABB

Computes the world-space AABB of this convex polyhedron, transformed by pos.

pub fn local_aabb(&self) -> AABB

Computes the local-space AABB of this convex polyhedron.

impl ConvexPolyhedron

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this convex polyhedron, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this convex polyhedron.

impl ConvexPolyhedron

pub fn from_convex_hull(points: &[Point<Real>]) -> Option<ConvexPolyhedron>

Creates a new convex polyhedron from an arbitrary set of points.

This explicitly computes the convex hull of the given set of points. Use Returns None if the convex hull computation failed.

pub fn from_convex_mesh(
points: Vec<Point<Real>>,
indices: &[[u32; 3]]
) -> Option<ConvexPolyhedron>

Attempts to create a new solid assumed to be convex from the set of points and indices.

The given points and index information are assumed to describe a convex polyhedron. It it is not, weird results may be produced.

Return

Retruns None if he given solid is not manifold (contains t-junctions, not closed, etc.)

pub fn check_geometry(&self)

Verify if this convex polyhedron is actually convex.

pub fn points(&self) -> &[Point<Real>]

The set of vertices of this convex polyhedron.

pub fn vertices(&self) -> &[Vertex]

The topology of the vertices of this convex polyhedron.

pub fn edges(&self) -> &[Edge]

The topology of the edges of this convex polyhedron.

pub fn faces(&self) -> &[Face]

The topology of the faces of this convex polyhedron.

pub fn vertices_adj_to_face(&self) -> &[u32]

The array containing the indices of the vertices adjacent to each face.

pub fn edges_adj_to_face(&self) -> &[u32]

The array containing the indices of the edges adjacent to each face.

pub fn faces_adj_to_vertex(&self) -> &[u32]

The array containing the indices of the faces adjacent to each vertex.

pub fn scaled(self, scale: &Vector<Real>) -> Option<Self>

Computes a scaled version of this convex polygon.

Returns None if the result had degenerate normals (for example if the scaling factor along one axis is zero).

pub fn support_feature_id_toward(
&self,
local_dir: &Unit<Vector<Real>>
) -> FeatureId

Computes the ID of the features with a normal that maximize the dot-product with local_dir.

pub fn feature_normal(&self, feature: FeatureId) -> Option<Unit<Vector<Real>>>

The normal of the given feature.

impl ConvexPolyhedron

pub fn to_trimesh(&self) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this convex polyhedron as a triangle-mesh.