Struct parry3d::shape::Cuboid

#[repr(C)]
pub struct Cuboid {
    pub half_extents: Vector<Real>,
}

Shape of a box.

Fields
half_extents: Vector<Real>

The half-extents of the cuboid.

Implementations

impl Cuboid

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Computes the world-space Aabb of this cuboid, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this cuboid.

impl Cuboid

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this cuboid, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this cuboid.

impl Cuboid

pub fn new(half_extents: Vector<Real>) -> Cuboid

Creates a new box from its half-extents. Half-extents are the box half-width along each axis. Each half-extent must be positive.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Computes a scaled version of this cuboid.

pub fn support_feature(&self, local_dir: Vector<Real>) -> PolygonalFeature

Return the face of this cuboid with a normal that maximizes the dot product with local_dir.

pub fn local_support_edge_segment(&self, local_dir: Vector<Real>) -> Segment

Return the edge segment of this cuboid with a normal cone containing a direction that that maximizes the dot product with local_dir.

pub fn support_face(&self, local_dir: Vector<Real>) -> PolygonalFeature

Computes the face with a normal that maximizes the dot-product with local_dir.

pub fn feature_normal(&self, feature: FeatureId) -> Option<Unit<Vector<Real>>>

The normal of the given feature of this shape.

impl Cuboid

pub fn to_outline(&self) -> (Vec<Point<Real>>, Vec<[u32; 2]>)

Outlines this cuboid’s shape using polylines.

impl Cuboid

pub fn to_trimesh(&self) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this cuboid as a triangle-mesh.