Struct parry3d::shape::Cylinder

#[repr(C)]
pub struct Cylinder {
    pub half_height: Real,
    pub radius: Real,
}

Cylinder shape with its principal axis aligned with the y axis.

Fields
half_height: Real

The half-height of the cylinder.

radius: Real

The radius fo the cylinder.

Implementations

impl Cylinder

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Computes the world-space Aabb of this cylinder, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this cylinder.

impl Cylinder

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this cylinder, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this cylinder.

impl Cylinder

pub fn new(half_height: Real, radius: Real) -> Cylinder

Creates a new cylinder.

Arguments:

half_height - the half length of the cylinder along the y axis.

radius - the length of the cylinder along all other axis.

pub fn scaled( self, scale: &Vector<Real>, nsubdivs: u32 ) -> Option<Either<Self, ConvexPolyhedron>>

Computes a scaled version of this cylinder.

If the scaling factor is non-uniform, then it can’t be represented as cylinder. Instead, a convex polyhedral approximation (with nsubdivs subdivisions) is returned. Returns None if that approximation had degenerate normals (for example if the scaling factor along one axis is zero).

impl Cylinder

pub fn to_outline(&self, nsubdiv: u32) -> (Vec<Point3<Real>>, Vec<[u32; 2]>)

Outlines this cylinder’s shape using polylines.

impl Cylinder

pub fn to_trimesh(&self, nsubdiv: u32) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this cylinder as a triangle-mesh.