Struct parry3d::shape::GenericHeightField

#[repr(C)]
pub struct GenericHeightField<Heights, Status> { /* private fields */ }

A 3D heightfield with a generic storage buffer for its height grid.

Implementations

impl<Heights, Status> GenericHeightField<Heights, Status> where
Heights: HeightFieldStorage<Item = Real>,
Status: HeightFieldStorage<Item = HeightFieldCellStatus>,

pub fn aabb(&self, pos: &Isometry<Real>) -> AABB

Computes the world-space AABB of this heightfield, transformed by pos.

pub fn local_aabb(&self) -> AABB

Computes the local-space AABB of this heightfield.

impl<Heights, Status> GenericHeightField<Heights, Status> where
Heights: HeightFieldStorage<Item = Real>,
Status: HeightFieldStorage<Item = HeightFieldCellStatus>,

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this height-field, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this height-field.

impl GenericHeightField<Matrix<f32, Dynamic, Dynamic, VecStorage<f32, Dynamic, Dynamic>>, Matrix<HeightFieldCellStatus, Dynamic, Dynamic, VecStorage<HeightFieldCellStatus, Dynamic, Dynamic>>>

pub fn new(heights: DMatrix<Real>, scale: Vector<Real>) -> Self

Initializes a new heightfield with the given heights and a scaling factor.

impl<Heights, Status> GenericHeightField<Heights, Status> where
Heights: HeightFieldStorage<Item = Real>,
Status: HeightFieldStorage<Item = HeightFieldCellStatus>,

pub fn nrows(&self) -> usize

The number of rows of this heightfield.

pub fn ncols(&self) -> usize

The number of columns of this heightfield.

pub fn closest_cell_at_point(&self, pt: &Point3<Real>) -> (usize, usize)

The pair of index of the cell containing the vertical projection of the given point.

pub fn cell_at_point(&self, pt: &Point3<Real>) -> Option<(usize, usize)>

The pair of index of the cell containing the vertical projection of the given point.

pub fn x_at(&self, j: usize) -> Real

The smallest x coordinate of the j-th column of this heightfield.

pub fn z_at(&self, i: usize) -> Real

The smallest z coordinate of the start of the i-th row of this heightfield.

pub fn triangles<'a>(&'a self) -> impl Iterator<Item = Triangle> + 'a

An iterator through all the triangles of this heightfield.

pub fn triangles_around_point<'a>(
&'a self,
point: &Point3<Real>
) -> HeightFieldRadialTriangles<'_, Heights, Status>

An iterator through all the triangles around the given point, after vertical projection on the heightfield.

pub fn triangles_at(
&self,
i: usize,
j: usize
) -> (Option<Triangle>, Option<Triangle>)

The two triangles at the cell (i, j) of this heightfield.

Returns None fore triangles that have been removed because of their user-defined status flags (described by the HeightFieldCellStatus bitfield).

pub fn cell_status(&self, i: usize, j: usize) -> HeightFieldCellStatus

The status of the (i, j)-th cell.

pub fn set_cell_status(
&mut self,
i: usize,
j: usize,
status: HeightFieldCellStatus
)

Set the status of the (i, j)-th cell.

pub fn cells_statuses(&self) -> &Status

The statuses of all the cells of this heightfield.

pub fn cells_statuses_mut(&mut self) -> &mut Status

The mutable statuses of all the cells of this heightfield.

pub fn heights(&self) -> &Heights

The heights of this heightfield.

pub fn scale(&self) -> &Vector<Real>

The scale factor applied to this heightfield.

pub fn set_scale(&mut self, new_scale: Vector<Real>)

Sets the scale factor applied to this heightfield.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Returns a scaled version of this heightfield.

pub fn cell_width(&self) -> Real

The width (extent along its local x axis) of each cell of this heightmap, including the scale factor.

pub fn cell_height(&self) -> Real

The height (extent along its local z axis) of each cell of this heightmap, including the scale factor.

pub fn unit_cell_width(&self) -> Real

The width (extent along its local x axis) of each cell of this heightmap, excluding the scale factor.

pub fn unit_cell_height(&self) -> Real

The height (extent along its local z axis) of each cell of this heightmap, excluding the scale factor.

pub fn root_aabb(&self) -> &AABB

The AABB of this heightmap.

pub fn convert_triangle_feature_id(
&self,
i: usize,
j: usize,
left: bool,
fid: FeatureId
) -> FeatureId

Converts the FeatureID of the left or right triangle at the cell (i, j) into a FeatureId of the whole heightfield.

pub fn map_elements_in_local_aabb(
&self,
aabb: &AABB,
f: &mut impl FnMut(u32, &Triangle)
)

Applies the function f to all the triangles of this heightfield intersecting the given AABB.

impl<Heights, Status> GenericHeightField<Heights, Status> where
Heights: HeightFieldStorage<Item = Real>,
Status: HeightFieldStorage<Item = HeightFieldCellStatus>,

pub fn to_trimesh(&self) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this ball as a triangle-mesh.