Struct parry3d::shape::GenericTriMesh

#[repr(C)]
pub struct GenericTriMesh<Storage: TriMeshStorage> { /* private fields */ }

A triangle mesh.

Implementations

impl GenericTriMesh<DefaultStorage>

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this triangle mesh, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this triangle mesh.

impl GenericTriMesh<DefaultStorage>

pub fn canonical_split( &self, axis: usize, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this TriMesh along the given canonical axis.

This will split the Aabb by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Result

Returns the result of the split. The first mesh returned is the piece lying on the negative half-space delimited by the splitting plane. The second mesh returned is the piece lying on the positive half-space delimited by the splitting plane.

pub fn split( &self, position: &Isometry<Real>, axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this mesh, transformed by position by a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn local_split( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this mesh by a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn canonical_intersection_with_plane( &self, axis: usize, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh and the plane identified by the given canonical axis.

This will intersect the mesh by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Note that the resultant polyline may have multiple connected components

pub fn intersection_with_plane( &self, position: &Isometry<Real>, axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh, transformed by position, and a plane identified by its normal axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn intersection_with_local_plane( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh and a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn intersection_with_aabb( &self, position: &Isometry<Real>, flip_mesh: bool, aabb: &Aabb, flip_cuboid: bool, epsilon: Real ) -> Option<Self>

Computes the intersection mesh between an Aabb and this mesh.

pub fn intersection_with_cuboid( &self, position: &Isometry<Real>, flip_mesh: bool, cuboid: &Cuboid, cuboid_position: &Isometry<Real>, flip_cuboid: bool, epsilon: Real ) -> Option<Self>

Computes the intersection mesh between a cuboid and this mesh transformed by position.

pub fn intersection_with_local_cuboid( &self, flip_mesh: bool, cuboid: &Cuboid, cuboid_position: &Isometry<Real>, flip_cuboid: bool, _epsilon: Real ) -> Option<Self>

Computes the intersection mesh between a cuboid and this mesh.

impl GenericTriMesh<DefaultStorage>

pub fn new(vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer.

pub fn with_flags( vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>, flags: TriMeshFlags ) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer, and flags controlling optional properties.

pub fn set_flags(&mut self, flags: TriMeshFlags) -> Result<(), TopologyError>

Sets the flags of this triangle mesh, controlling its optional associated data.

pub fn transform_vertices(&mut self, transform: &Isometry<Real>)

Transforms in-place the vertices of this triangle mesh.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Returns a scaled version of this triangle mesh.

pub fn append(&mut self, rhs: &TriMesh)

Appends a second triangle mesh to this triangle mesh.

pub fn flat_indices(&self) -> &[u32]

A flat view of the index buffer of this mesh.

pub fn reverse(&mut self)

Reverse the orientation of the triangle mesh.

pub fn triangles(&self) -> impl ExactSizeIterator<Item = Triangle> + '_

An iterator through all the triangles of this mesh.

impl<Storage: TriMeshStorage> GenericTriMesh<Storage>

pub fn flags(&self) -> TriMeshFlags

The flags of this triangle mesh.

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Compute the axis-aligned bounding box of this triangle mesh.

pub fn local_aabb(&self) -> &Aabb

Gets the local axis-aligned bounding box of this triangle mesh.

pub fn qbvh(&self) -> &GenericQbvh<u32, Storage::QbvhStorage>

The acceleration structure used by this triangle-mesh.

pub fn num_triangles(&self) -> usize

The number of triangles forming this mesh.

pub fn is_backface(&self, feature: FeatureId) -> bool

Does the given feature ID identify a backface of this trimesh?

pub fn triangle(&self, i: u32) -> Triangle

Get the i-th triangle of this mesh.

pub fn vertices(&self) -> &Storage::ArrayPoint

The vertex buffer of this mesh.

pub fn indices(&self) -> &Storage::ArrayIdx

The index buffer of this mesh.

pub fn topology(&self) -> Option<&TriMeshTopology<Storage>>

Returns the topology information of this trimesh, if it has been computed.

pub fn connected_components( &self ) -> Option<&TriMeshConnectedComponents<Storage>>

Returns the connected-component information of this trimesh, if it has been computed.

pub fn pseudo_normals(&self) -> Option<&TriMeshPseudoNormals<Storage>>

The pseudo-normals of this triangle mesh, if they have been computed.