Struct parry3d::shape::Segment

#[repr(C)]
pub struct Segment {
    pub a: Point<Real>,
    pub b: Point<Real>,
}

A segment shape.

Fields
a: Point<Real>

The segment first point.

b: Point<Real>

The segment second point.

Implementations

impl Segment

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Computes the world-space Aabb of this segment, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this segment.

impl Segment

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this segment, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this segment.

impl Segment

pub fn canonical_split( &self, axis: usize, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this segment along the given canonical axis.

This will split the segment by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Result

Returns the result of the split. The first shape returned is the piece lying on the negative half-space delimited by the splitting plane. The second shape returned is the piece lying on the positive half-space delimited by the splitting plane.

pub fn local_split( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this segment by a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn local_split_and_get_intersection( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> (SplitResult<Self>, Option<(Point<Real>, Real)>)

Split a segment with a plane.

This returns the result of the splitting operation, as well as the intersection point (and barycentric coordinate of this point) with the plane. The intersection point is None if the plane is parallel or near-parallel to the segment.

impl Segment

pub fn new(a: Point<Real>, b: Point<Real>) -> Segment

Creates a new segment from two points.

pub fn from_array(arr: &[Point<Real>; 2]) -> &Segment

Creates the reference to a segment from the reference to an array of two points.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Computes a scaled version of this segment.

pub fn scaled_direction(&self) -> Vector<Real>

The direction of this segment scaled by its length.

Points from self.a toward self.b.

pub fn length(&self) -> Real

The length of this segment.

pub fn swap(&mut self)

Swaps the two vertices of this segment.

pub fn direction(&self) -> Option<Unit<Vector<Real>>>

The unit direction of this segment.

Points from self.a() toward self.b(). Returns None is both points are equal.

pub fn scaled_planar_normal(&self, plane_axis: u8) -> Vector<Real>

The not-normalized counterclockwise normal of this segment, assuming it lies on the plane with the normal collinear to the given axis (0 = X, 1 = Y, 2 = Z).

pub fn normal(&self) -> Option<Unit<Vector<Real>>>

Returns None. Exists only for API similarity with the 2D parry.

pub fn planar_normal(&self, plane_axis: u8) -> Option<Unit<Vector<Real>>>

The normalized counterclockwise normal of this segment, assuming it lies on the plane with the normal collinear to the given axis (0 = X, 1 = Y, 2 = Z).

pub fn transformed(&self, m: &Isometry<Real>) -> Self

Applies the isometry m to the vertices of this segment and returns the resulting segment.

pub fn point_at(&self, location: &SegmentPointLocation) -> Point<Real>

Computes the point at the given location.

pub fn feature_normal(&self, feature: FeatureId) -> Option<Unit<Vector<Real>>>

The normal of the given feature of this shape.