Struct parry3d::shape::SharedShape

pub struct SharedShape(pub Arc<dyn Shape>);

The shape of a collider.

Tuple Fields
0: Arc<dyn Shape>
Implementations

impl SharedShape

pub fn new(shape: impl Shape) -> Self

Wraps the given shape as a shared shape.

pub fn make_mut(&mut self) -> &mut dyn Shape

If this shape is shared, then the content of self is cloned into a unique instance, and a mutable reference to that instance is returned.

pub fn compound(shapes: Vec<(Isometry<Real>, SharedShape)>) -> Self

Initialize a compound shape defined by its subshapes.

pub fn ball(radius: Real) -> Self

Initialize a ball shape defined by its radius.

pub fn halfspace(outward_normal: Unit<Vector<Real>>) -> Self

Initialize a plane shape defined by its outward normal.

pub fn cylinder(half_height: Real, radius: Real) -> Self

Initialize a cylindrical shape defined by its half-height (along along the y axis) and its radius.

pub fn round_cylinder(
half_height: Real,
radius: Real,
border_radius: Real
) -> Self

Initialize a rounded cylindrical shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn round_cone(half_height: Real, radius: Real, border_radius: Real) -> Self

Initialize a rounded cone shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn cone(half_height: Real, radius: Real) -> Self

Initialize a cone shape defined by its half-height (along along the y axis) and its basis radius.

pub fn cuboid(hx: Real, hy: Real, hz: Real) -> Self

Initialize a cuboid shape defined by its half-extents.

pub fn round_cuboid(hx: Real, hy: Real, hz: Real, border_radius: Real) -> Self

Initialize a round cuboid shape defined by its half-extents and border radius.

pub fn capsule(a: Point<Real>, b: Point<Real>, radius: Real) -> Self

Initialize a capsule shape from its endpoints and radius.

pub fn capsule_x(half_height: Real, radius: Real) -> Self

Initialize a capsule shape aligned with the x axis.

pub fn capsule_y(half_height: Real, radius: Real) -> Self

Initialize a capsule shape aligned with the y axis.

pub fn capsule_z(half_height: Real, radius: Real) -> Self

Initialize a capsule shape aligned with the z axis.

pub fn segment(a: Point<Real>, b: Point<Real>) -> Self

Initialize a segment shape from its endpoints.

pub fn triangle(a: Point<Real>, b: Point<Real>, c: Point<Real>) -> Self

Initializes a triangle shape.

pub fn round_triangle(
a: Point<Real>,
b: Point<Real>,
c: Point<Real>,
border_radius: Real
) -> Self

Initializes a triangle shape with round corners.

pub fn polyline(
vertices: Vec<Point<Real>>,
indices: Option<Vec<[u32; 2]>>
) -> Self

Initializes a polyline shape defined by its vertex and index buffers.

If no index buffer is provided, the polyline is assumed to describe a line strip.

pub fn trimesh(vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>) -> Self

Initializes a triangle mesh shape defined by its vertex and index buffers.

pub fn convex_decomposition(
vertices: &[Point<Real>],
indices: &[[u32; 3]]
) -> Self

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
border_radius: Real
) -> Self

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_decomposition_with_params(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
params: &VHACDParameters
) -> Self

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition_with_params(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
params: &VHACDParameters,
border_radius: Real
) -> Self

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_hull(points: &[Point<Real>]) -> Option<Self>

Creates a new shared shape that is the convex-hull of the given points.

pub fn convex_mesh(
points: Vec<Point<Real>>,
indices: &[[u32; 3]]
) -> Option<Self>

Creates a new shared shape that is a convex polyhedron formed by the given set of points assumed to form a convex mesh (no convex-hull will be automatically computed).

pub fn round_convex_hull(
points: &[Point<Real>],
border_radius: Real
) -> Option<Self>

Creates a new shared shape with rounded corners that is the convex-hull of the given points, dilated by border_radius.

pub fn round_convex_mesh(
points: Vec<Point<Real>>,
indices: &[[u32; 3]],
border_radius: Real
) -> Option<Self>

Creates a new shared shape with round corners that is a convex polyhedron formed by the given set of points assumed to form a convex mesh (no convex-hull will be automatically computed).

pub fn heightfield(heights: DMatrix<Real>, scale: Vector<Real>) -> Self

Initializes an heightfield shape on the x-z plane defined by its set of height and a scale factor along each coordinate axis.