Trait parry3d::shape::PolygonalFeatureMap

pub trait PolygonalFeatureMap: SupportMap {
    fn local_support_feature(
        &self, 
        dir: &Unit<Vector<Real>>, 
        out_feature: &mut PolygonalFeature
    );
}

Trait implemented by convex shapes with features with polyhedral approximations.

Required Methods

fn local_support_feature(
&self,
dir: &Unit<Vector<Real>>,
out_feature: &mut PolygonalFeature
)

Compute the support polygonal face of self towards the dir.