Trait parry3d::shape::SimdCompositeShape

pub trait SimdCompositeShape {
    // Required methods
    fn map_part_at(
        &self,
        shape_id: u32,
        f: &mut dyn FnMut(Option<&Isometry<Real>>, &dyn Shape)
    );
    fn qbvh(&self) -> &Qbvh<u32>;
}

Trait implemented by shapes composed of multiple simpler shapes.

A composite shape is composed of several shapes. For example, this can be a convex decomposition of a concave shape; or a triangle-mesh.

Required Methods

fn map_part_at( &self, shape_id: u32, f: &mut dyn FnMut(Option<&Isometry<Real>>, &dyn Shape) )

Applies a function to one sub-shape of this composite shape.

fn qbvh(&self) -> &Qbvh<u32>

Gets the acceleration structure of the composite shape.