Trait parry3d::shape::SupportMap

pub trait SupportMap {
    fn local_support_point(&self, dir: &Vector<Real>) -> Point<Real>;

    fn local_support_point_toward(
        &self, 
        dir: &Unit<Vector<Real>>
    ) -> Point<Real> { ... }
    fn support_point(
        &self, 
        transform: &Isometry<Real>, 
        dir: &Vector<Real>
    ) -> Point<Real> { ... }
    fn support_point_toward(
        &self, 
        transform: &Isometry<Real>, 
        dir: &Unit<Vector<Real>>
    ) -> Point<Real> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required Methods

fn local_support_point(&self, dir: &Vector<Real>) -> Point<Real>

Provided Methods

fn local_support_point_toward(&self, dir: &Unit<Vector<Real>>) -> Point<Real>

Same as self.local_support_point except that dir is normalized.

fn support_point(
&self,
transform: &Isometry<Real>,
dir: &Vector<Real>
) -> Point<Real>

fn support_point_toward(
&self,
transform: &Isometry<Real>,
dir: &Unit<Vector<Real>>
) -> Point<Real>

Same as self.support_point except that dir is normalized.