Type Alias parry3d::shape::HeightField

pub type HeightField = GenericHeightField<DefaultStorage>;

A 3D heightfield.

Aliased Type

struct HeightField { /* private fields */ }

Implementations

impl HeightField

pub fn new(heights: DMatrix<Real>, scale: Vector<Real>) -> Self

Initializes a new heightfield with the given heights and a scaling factor.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Computes the world-space Aabb of this heightfield, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this heightfield.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this height-field, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this height-field.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn nrows(&self) -> usize

The number of rows of this heightfield.

pub fn ncols(&self) -> usize

The number of columns of this heightfield.

pub fn closest_cell_at_point(&self, pt: &Point3<Real>) -> (usize, usize)

The pair of index of the cell containing the vertical projection of the given point.

pub fn cell_at_point(&self, pt: &Point3<Real>) -> Option<(usize, usize)>

The pair of index of the cell containing the vertical projection of the given point.

pub fn unclamped_cell_at_point(&self, pt: &Point3<Real>) -> (isize, isize)

The pair of index of the cell containing the vertical projection of the given point.

pub fn x_at(&self, j: usize) -> Real

The smallest x coordinate of the j-th column of this heightfield.

pub fn z_at(&self, i: usize) -> Real

The smallest z coordinate of the start of the i-th row of this heightfield.

pub fn signed_x_at(&self, j: isize) -> Real

The smallest x coordinate of the j-th column of this heightfield.

pub fn signed_z_at(&self, i: isize) -> Real

The smallest z coordinate of the start of the i-th row of this heightfield.

pub fn triangles<'a>(&'a self) -> impl Iterator<Item = Triangle> + 'a

An iterator through all the triangles of this heightfield.

pub fn triangles_around_point<'a>( &'a self, point: &Point3<Real> ) -> HeightFieldRadialTriangles<'_, Storage>

An iterator through all the triangles around the given point, after vertical projection on the heightfield.

pub fn triangle_at_id(&self, id: u32) -> Option<Triangle>

Gets the the vertices of the triangle identified by id.

pub fn triangle_vids_at_id(&self, id: u32) -> Option<[u32; 3]>

Gets the vertex indices of the triangle identified by id.

pub fn triangles_vids_at( &self, i: usize, j: usize ) -> (Option<[u32; 3]>, Option<[u32; 3]>)

Gets the indices of the vertices of the (up to) two triangles for the cell (i, j).

pub fn triangles_at( &self, i: usize, j: usize ) -> (Option<Triangle>, Option<Triangle>)

The two triangles at the cell (i, j) of this heightfield.

Returns None fore triangles that have been removed because of their user-defined status flags (described by the HeightFieldCellStatus bitfield).

pub fn num_cells_ij(&self) -> (usize, usize)

The number of cells of this heightfield along each dimension.

pub fn cell_status(&self, i: usize, j: usize) -> HeightFieldCellStatus

The status of the (i, j)-th cell.

pub fn set_cell_status( &mut self, i: usize, j: usize, status: HeightFieldCellStatus )

Set the status of the (i, j)-th cell.

pub fn cells_statuses(&self) -> &Storage::Status

The statuses of all the cells of this heightfield.

pub fn cells_statuses_mut(&mut self) -> &mut Storage::Status

The mutable statuses of all the cells of this heightfield.

pub fn heights(&self) -> &Storage::Heights

The heights of this heightfield.

pub fn scale(&self) -> &Vector<Real>

The scale factor applied to this heightfield.

pub fn set_scale(&mut self, new_scale: Vector<Real>)

Sets the scale factor applied to this heightfield.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Returns a scaled version of this heightfield.

pub fn cell_width(&self) -> Real

The width (extent along its local x axis) of each cell of this heightmap, including the scale factor.

pub fn cell_height(&self) -> Real

The height (extent along its local z axis) of each cell of this heightmap, including the scale factor.

pub fn unit_cell_width(&self) -> Real

The width (extent along its local x axis) of each cell of this heightmap, excluding the scale factor.

pub fn unit_cell_height(&self) -> Real

The height (extent along its local z axis) of each cell of this heightmap, excluding the scale factor.

pub fn root_aabb(&self) -> &Aabb

The Aabb of this heightmap.

pub fn convert_triangle_feature_id( &self, i: usize, j: usize, left: bool, fid: FeatureId ) -> FeatureId

Converts the FeatureID of the left or right triangle at the cell (i, j) into a FeatureId of the whole heightfield.

pub fn unclamped_elements_range_in_local_aabb( &self, aabb: &Aabb ) -> (Range<isize>, Range<isize>)

The range of segment ids that may intersect the given local Aabb.

pub fn map_elements_in_local_aabb( &self, aabb: &Aabb, f: &mut impl FnMut(u32, &Triangle) )

Applies the function f to all the triangles of this heightfield intersecting the given Aabb.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn to_trimesh(&self) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this heightfield as a triangle-mesh.