Type Alias parry3d::shape::TriMesh

pub type TriMesh = GenericTriMesh<DefaultStorage>;

A triangle-mesh.

Aliased Type

struct TriMesh { /* private fields */ }

Implementations

impl TriMesh

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this triangle mesh, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this triangle mesh.

impl TriMesh

pub fn canonical_split( &self, axis: usize, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this TriMesh along the given canonical axis.

This will split the Aabb by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Result

Returns the result of the split. The first mesh returned is the piece lying on the negative half-space delimited by the splitting plane. The second mesh returned is the piece lying on the positive half-space delimited by the splitting plane.

pub fn split( &self, position: &Isometry<Real>, axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this mesh, transformed by position by a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn local_split( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this mesh by a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn canonical_intersection_with_plane( &self, axis: usize, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh and the plane identified by the given canonical axis.

This will intersect the mesh by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Note that the resultant polyline may have multiple connected components

pub fn intersection_with_plane( &self, position: &Isometry<Real>, axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh, transformed by position, and a plane identified by its normal axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn intersection_with_local_plane( &self, local_axis: &UnitVector<Real>, bias: Real, epsilon: Real ) -> IntersectResult<Polyline>

Computes the intersection Polylines between this mesh and a plane identified by its normal local_axis and the bias (i.e. the plane passes through the point equal to normal * bias).

pub fn intersection_with_aabb( &self, position: &Isometry<Real>, flip_mesh: bool, aabb: &Aabb, flip_cuboid: bool, epsilon: Real ) -> Option<Self>

Computes the intersection mesh between an Aabb and this mesh.

pub fn intersection_with_cuboid( &self, position: &Isometry<Real>, flip_mesh: bool, cuboid: &Cuboid, cuboid_position: &Isometry<Real>, flip_cuboid: bool, epsilon: Real ) -> Option<Self>

Computes the intersection mesh between a cuboid and this mesh transformed by position.

pub fn intersection_with_local_cuboid( &self, flip_mesh: bool, cuboid: &Cuboid, cuboid_position: &Isometry<Real>, flip_cuboid: bool, _epsilon: Real ) -> Option<Self>

Computes the intersection mesh between a cuboid and this mesh.

impl TriMesh

pub fn new(vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer.

pub fn with_flags( vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>, flags: TriMeshFlags ) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer, and flags controlling optional properties.

pub fn set_flags(&mut self, flags: TriMeshFlags) -> Result<(), TopologyError>

Sets the flags of this triangle mesh, controlling its optional associated data.

pub fn transform_vertices(&mut self, transform: &Isometry<Real>)

Transforms in-place the vertices of this triangle mesh.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Returns a scaled version of this triangle mesh.

pub fn append(&mut self, rhs: &TriMesh)

Appends a second triangle mesh to this triangle mesh.

pub fn flat_indices(&self) -> &[u32]

A flat view of the index buffer of this mesh.

pub fn reverse(&mut self)

Reverse the orientation of the triangle mesh.

pub fn triangles(&self) -> impl ExactSizeIterator<Item = Triangle> + '_

An iterator through all the triangles of this mesh.