Module parry3d::transformation

Transformation, simplification and decomposition of meshes.

Modules

utils

Utilities useful for various generations tasks.

vhacd

Approximate convex decomposition using the VHACD algorithm.

voxelization

Voxelization of a 2D polyline or 3D triangle mesh.

Enums

ConvexHullError

Errors generated by the convex-hull calculation.

Functions

check_convex_hull

Checks if a convex-hull is properly formed.

convex_hull

Computes the convex hull of a set of 3d points.

convex_polygons_intersection

Computes the intersection of two convex polygons.

convex_polygons_intersection_points

Computes the intersection points of two convex polygons.

intersect_meshes

Computes the intersection of two meshes.

try_convex_hull

Computes the convex hull of a set of 3d points.