Module parry3d::utils::hashmap

A hash-map that behaves deterministically when the enhanced-determinism feature is enabled.

Structs

FxHasher32

This is the same as FxHasher, but with the guarantee that the internal hash is an u32 instead of something that depends on the platform.

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.

Type Definitions

HashMap

Type alias for a hashmap using the fx hash algorithm.