Struct parry3d::bounding_volume::Aabb

pub struct Aabb {
    pub mins: Point<Real>,
    pub maxs: Point<Real>,
}

An Axis Aligned Bounding Box.

Fields
mins: Point<Real>maxs: Point<Real>
Implementations

impl Aabb

pub const EDGES_VERTEX_IDS: [(usize, usize); 12] = _

The vertex indices of each edge of this Aabb.

This gives, for each edge of this Aabb, the indices of its vertices when taken from the self.vertices() array. Here is how the faces are numbered, assuming a right-handed coordinate system:

y 3 - 2 | 7 − 6 | ___ x | | 1 (the zero is bellow 3 and on the left of 1, hidden by the 4-5-6-7 face.) / 4 - 5 z

pub const FACES_VERTEX_IDS: [(usize, usize, usize, usize); 6] = _

The vertex indices of each face of this Aabb.

This gives, for each face of this Aabb, the indices of its vertices when taken from the self.vertices() array. Here is how the faces are numbered, assuming a right-handed coordinate system:

y 3 - 2 | 7 − 6 | ___ x | | 1 (the zero is bellow 3 and on the left of 1, hidden by the 4-5-6-7 face.) / 4 - 5 z

pub fn new(mins: Point<Real>, maxs: Point<Real>) -> Aabb

Creates a new Aabb.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

pub fn new_invalid() -> Self

Creates an invalid Aabb with mins components set to Real::max_values and maxscomponents set to -Real::max_values.

This is often used as the initial values of some Aabb merging algorithms.

pub fn from_half_extents( center: Point<Real>, half_extents: Vector<Real> ) -> Self

Creates a new Aabb from its center and its half-extents.

pub fn from_points<'a, I>(pts: I) -> Selfwhere I: IntoIterator<Item = &'a Point<Real>>,

Creates a new Aabb from a set of points.

pub fn center(&self) -> Point<Real>

The center of this Aabb.

pub fn half_extents(&self) -> Vector<Real>

The half extents of this Aabb.

pub fn volume(&self) -> Real

The volume of this Aabb.

pub fn extents(&self) -> Vector<Real>

The extents of this Aabb.

pub fn take_point(&mut self, pt: Point<Real>)

Enlarges this Aabb so it also contains the point pt.

pub fn transform_by(&self, m: &Isometry<Real>) -> Self

Computes the Aabb bounding self transformed by m.

pub fn scaled(self, scale: &Vector<Real>) -> Self

pub fn bounding_sphere(&self) -> BoundingSphere

The smallest bounding sphere containing this Aabb.

pub fn contains_local_point(&self, point: &Point<Real>) -> bool

pub fn intersection(&self, other: &Aabb) -> Option<Aabb>

Computes the intersection of this Aabb and another one.

pub fn difference(&self, rhs: &Aabb) -> ArrayVec<Self, TWO_DIM>

Returns the difference between this Aabb and rhs.

Removing another Aabb from self will result in zero, one, or up to 4 (in 2D) or 8 (in 3D) new smaller Aabbs.

pub fn difference_with_cut_sequence( &self, rhs: &Aabb ) -> (ArrayVec<Self, TWO_DIM>, ArrayVec<(i8, Real), TWO_DIM>)

Returns the difference between this Aabb and rhs.

Removing another Aabb from self will result in zero, one, or up to 4 (in 2D) or 8 (in 3D) new smaller Aabbs.

Return

This returns a pair where the first item are the new Aabbs and the the second item is the sequance of cuts applied to self to obtain the new Aabbs. Each cut is performed along one axis identified by -1, -2, -3 for -X, -Y, -Z and 1, 2, 3 for +X, +Y, +Z, and the plane’s bias. The cuts are applied sequancially. For example, if result.1[0] contains 1, then it means that result.0[0] is equal to the piece of self lying in the negative half-space delimited by the plane with outward normal +X. Then, the other piece of self generated by this cut (i.e. the piece of self lying in the positive half-space delimited by the plane with outward normal +X) is the one that will be affected by the next cut.

The returned cut sequence will be empty if the aabbs are disjoint.

pub fn vertices(&self) -> [Point<Real>; 8]

Computes the vertices of this Aabb.

pub fn split_at_center(&self) -> [Aabb; 8]

Splits this Aabb at its center, into height parts (as in an octree).

pub fn project_on_axis(&self, axis: &UnitVector<Real>) -> (Real, Real)

Projects every point of Aabb on an arbitrary axis.

pub fn intersects_spiral( &self, point: &Point<Real>, center: &Point<Real>, axis: &UnitVector<Real>, linvel: &Vector<Real>, angvel: Real ) -> bool

impl Aabb

pub fn clip_segment( &self, pa: &Point<Real>, pb: &Point<Real> ) -> Option<Segment>

Computes the intersection of a segment with this Aabb.

Returns None if there is no intersection.

pub fn clip_line_parameters( &self, orig: &Point<Real>, dir: &Vector<Real> ) -> Option<(Real, Real)>

Computes the parameters of the two intersection points between a line and this Aabb.

The parameters are such that the point are given by orig + dir * parameter. Returns None if there is no intersection.

pub fn clip_line( &self, orig: &Point<Real>, dir: &Vector<Real> ) -> Option<Segment>

Computes the intersection segment between a line and this Aabb.

Returns None if there is no intersection.

pub fn clip_ray_parameters(&self, ray: &Ray) -> Option<(Real, Real)>

Computes the parameters of the two intersection points between a ray and this Aabb.

The parameters are such that the point are given by ray.orig + ray.dir * parameter. Returns None if there is no intersection.

pub fn clip_ray(&self, ray: &Ray) -> Option<Segment>

Computes the intersection segment between a ray and this Aabb.

Returns None if there is no intersection.

impl Aabb

pub fn clip_polygon(&self, points: &mut Vec<Point<Real>>)

Computes the intersections between this Aabb and the given polygon.

The results is written into points directly. The input points are assumed to form a convex polygon where all points lie on the same plane. In order to avoid internal allocations, uses self.clip_polygon_with_workspace instead.

pub fn clip_polygon_with_workspace( &self, points: &mut Vec<Point<Real>>, workspace: &mut Vec<Point<Real>> )

Computes the intersections between this Aabb and the given polygon.

The results is written into points directly. The input points are assumed to form a convex polygon where all points lie on the same plane.

impl Aabb

pub fn canonical_split( &self, axis: usize, bias: Real, epsilon: Real ) -> SplitResult<Self>

Splits this Aabb along the given canonical axis.

This will split the Aabb by a plane with a normal with it’s axis-th component set to 1. The splitting plane is shifted wrt. the origin by the bias (i.e. it passes through the point equal to normal * bias).

Result

Returns the result of the split. The first Aabb returned is the piece lying on the negative half-space delimited by the splitting plane. The second Aabb returned is the piece lying on the positive half-space delimited by the splitting plane.

impl Aabb

pub fn to_outline(&self) -> (Vec<Point<Real>>, Vec<[u32; 2]>)

Outlines this Aabb’s shape using polylines.

impl Aabb

pub fn to_trimesh(&self) -> (Vec<Point3<Real>>, Vec<[u32; 3]>)

Discretize the boundary of this Aabb as a triangle-mesh.