Struct parry3d::transformation::voxelization::VoxelSet

pub struct VoxelSet {
    pub origin: Point<Real>,
    pub scale: Real,
    /* private fields */
}

A sparse set of voxels.

It only contains voxels that are considered as “full” after a voxelization.

Fields
origin: Point<Real>

The 3D origin of this voxel-set.

scale: Real

The scale factor between the voxel integer coordinates and their actual float world-space coordinates.

Implementations

impl VoxelSet

pub fn new() -> Self

Creates a new empty set of voxels.

pub fn voxel_volume(&self) -> Real

The volume of a single voxel of this voxel set.

pub fn voxelize( points: &[Point<Real>], indices: &[[u32; 3]], resolution: u32, fill_mode: FillMode, keep_voxel_to_primitives_map: bool ) -> Self

Voxelizes the given shape described by its boundary: a triangle mesh (in 3D) or polyline (in 2D).

Parameters

points - The vertex buffer of the boundary of the shape to voxelize.

indices - The index buffer of the boundary of the shape to voxelize.

resolution - Controls the number of subdivision done along each axis. This number is the number of subdivisions along the axis where the input shape has the largest extent. The other dimensions will have a different automatically-determined resolution (in order to keep the voxels cubic).

fill_mode - Controls what is being voxelized.

keep_voxel_to_primitives_map - If set to true a map between the voxels and the primitives (3D triangles or 2D segments) it intersects will be computed.

pub fn min_bb_voxels(&self) -> Point<u32>

The minimal coordinates of the integer bounding-box of the voxels in this set.

pub fn max_bb_voxels(&self) -> Point<u32>

The maximal coordinates of the integer bounding-box of the voxels in this set.

pub fn compute_volume(&self) -> Real

Computes the total volume of the voxels contained by this set.

pub fn len(&self) -> usize

The number of voxels in this set.

pub fn voxels(&self) -> &[Voxel]

The set of voxels.

pub fn compute_bb(&mut self)

Update the bounding box of this voxel set.

pub fn compute_exact_convex_hull( &self, points: &[Point<Real>], indices: &[[u32; 3]] ) -> (Vec<Point<Real>>, Vec<[u32; 3]>)

Compute the convex-hull of this voxel set after cutting each voxel by the primitives (3D triangle or 2D segments) it intersects.

This will panic if this VoxelSet was created with keep_voxel_to_primitives_map = false.

pub fn compute_primitive_intersections( &self, points: &[Point<Real>], indices: &[[u32; 3]] ) -> Vec<Point<Real>>

Computes the intersections between all the voxels of this voxel set, and all the primitives (triangle or segments) it intersected (as per the voxel-to-primitives-map computed during voxelization).

Panics if the voxelization was performed without setting the parameter voxel_to_primitives_map = true.

pub fn compute_convex_hull( &self, sampling: u32 ) -> (Vec<Point<Real>>, Vec<[u32; 3]>)

Compute the convex-hull of the voxels in this set.

Parameters

sampling - The convex-hull computation will ignore sampling voxels at regular intervals. Useful to save some computation times if an exact result isn’t need. Use 0 to make sure no voxel is being ignored.

pub fn to_trimesh( &self, base_index: u32, is_on_surface: bool ) -> (Vec<Point<Real>>, Vec<[u32; 3]>)

Convert self into a mesh, including only the voxels on the surface or only the voxel inside of the volume.