Module parry3d::utils

Various unsorted geometrical and logical operators.

Modules

hashmap

A hash-map that behaves deterministically when the enhanced-determinism feature is enabled.

Structs

DefaultStorage

Default data storage based on Vec, DVector, and DMatrix.

DeterministicState

A hasher builder that creates DefaultHasher with default keys.

HashablePartialEq

A structure that implements Eq and is hashable even if the wrapped data implements only PartialEq.

Interval

An interval implementing interval arithmetic.

SdpMatrix2

A 2x2 symmetric-definite-positive matrix.

SdpMatrix3

A 3x3 symmetric-definite-positive matrix.

SortedPair

A pair of elements sorted in increasing order.

Enums

SegmentsIntersection

Intersection between two segments.

Traits

Array1

Abstraction over a 1D array.

Array2

Abstraction over a 2D array.

AsBytes

Trait that transforms thing to a slice of u8.

IntervalFunction

A derivable valued function which can be bounded on intervals.

IsometryOps

Extra operations with isometries.

IsometryOpt

Various operations usable with Option<Isometry> and Option<&Isometry> where None is assumed to be equivalent to the identity.

Functions

ccw_face_normal

Computes the normal of a counter-clock-wise triangle.

center

Computes the center of a set of point.

center_cov

Computes the center and the covariance matrix of a set of points.

cov

Computes the covariance matrix of a set of points.

find_root_intervals

Execute the Interval Newton Method to isolate all the roots of the given nonlinear function.

find_root_intervals_to

Execute the Interval Newton Method to isolate all the roots of the given nonlinear function.

median

Computes the median of a set of values.

obb

Computes an oriented bounding box for the given set of points.

point_cloud_support_point

Computes the support point of a cloud of points.

point_cloud_support_point_id

Computes the index of the support point of a cloud of points.

point_in_convex_poly2d

Tests if the given point is inside a convex polygon with arbitrary orientation.

point_in_poly2d

Tests if the given point is inside an arbitrary closed polygon with arbitrary orientation, using a counting winding strategy.

remove_unused_points

Given an index buffer, remove from points every point that is not indexed.

segments_intersection2d

Computes the intersection between two segments.