Struct parry3d::utils::Interval

pub struct Interval<T>(pub T, pub T);

An interval implementing interval arithmetic.

Tuple Fields
0: T1: T
Implementations

impl<T> Interval<T>

pub fn sort(a: T, b: T) -> Selfwhere T: PartialOrd,

Create the interval [min(a, b), max(a, b)].

pub fn splat(e: T) -> Selfwhere T: Clone,

Create the interval [e, e] (single value).

pub fn contains(&self, t: T) -> boolwhere T: PartialOrd<T>,

Does this interval contain the given value?

pub fn width(self) -> T::Outputwhere T: Sub<T>,

The width of this inverval.

pub fn midpoint(self) -> Twhere T: RealField + Copy,

The average of the two interval endpoints.

pub fn split(self) -> [Self; 2]where T: RealField + Copy,

Splits this interval at its mitpoint.

pub fn enclose(self, t: T) -> Selfwhere T: PartialOrd,

Computes a new interval that contains both self and t.

pub fn intersect(self, rhs: Self) -> Option<Self>where T: PartialOrd + SimdPartialOrd,

Computes the intersection between two intervals.

Returns None if the intervals are disjoint.

pub fn sin_cos(self) -> (Self, Self)where T: RealField + Copy,

Bounds the image of thesin and cos functions on this interval.

pub fn sin(self) -> Selfwhere T: RealField + Copy,

Bounds the image of the sinus function on this interval.

pub fn cos(self) -> Selfwhere T: RealField + Copy,

Bounds the image of the cosinus function on this interval.